requirements.txt
wheel
setuptools
twine
django
djangorestframework
django-structlog